Add viewport by default; set width/height if parameters are given
Also if viewport=false parameter is present, don't add anything

Original Commit Message from woneal1983:
I can do without the viewport tag all together on my amazon and google tv apps, but for the new chromium 43 library for some reason has to have the specific height and width in the viewport in order to fix a scaling issue on boxes and sticks. Works right on tablets but not on boxes or sticks. So if you don't want any viewport at all then this new pull request will do what is needed. I also changed the parameters to vh and vw which works better for me. Anyway take a look and see if this fits.